Retain meta from wrapper script during buildEnv
Currently, no metadata is set for the environment, which leads to
the wrong assumption that the main program used when running the
environment equals the package name. In our case this is the
original package name prefixed with "nixpak-".
However, the wrapper script is named like the original main
program, which leads to errors like the following when running
the environment:
  error: unable to execute '/nix/store/apnm35wckzlq4nqcmk3zpwys07srn9y4-nixpak-ungoogled-chromium-127.0.6533.99/bin/nixpak-ungoogled-chromium': No such file or directory

This commit retains the metadata from the wrapper script when
building the environment, which also copies meta.mainProgram.
Therefore, the correct main program is detected when running the
environment.