Refactor launcher into a sandbox supervisor
Previously the launcher just set up everything and passed the
execution on to bwrap. So the launcher was only running initially,
but not throughout the whole sandbox lifetime.
This was fine with just xdg-dbus-proxy, since bwrap could directly
communicate with it via pipes to coordinate shutdown.
However, with more sandboxing daemons upcoming, this complicates
correct sandbox setup and shutdown coordination.
Therefore, this commit refactors the launcher into a sandbox
supervisor to coordinate sandbox setup and shutdown and prepare to
accomodate different requirements of various sandboxing daemons.