Add support for processes that fork into background
The re-worked launcher killed the whole sandbox when the parent
bwrap process exited. This happens when the main process within the
sandbox exits. However, if the main process forked into background,
it is also killed.
Before the re-work the bwrap child process and background processes
kept running.
This commit explicitly waits for the bwrap child process to exit,
which is the case when all processes within the sandbox have
exited.
Since the bwrap child process is not a child process of our own,
but a child process of the bwrap parent process, we need to become
a child reaper for grandchildren.
The new behavior still slightly differs from before the launcher
re-work, since the launcher keeps running even when the main
process already exited, basically waiting and blocking until the
exit of the whole sandbox.