Fix sandbox termination when pasta fails
Previously, when pasta failed to setup the network namespace the
bwrap child process continued to run, which in turn also spawned
the application. Since the network namespace is always created by
bubblewrap, pasta usually leaves the network namespace as it is
(in a failsafe default with only a loopback interface) when it
fails.
With this commit we explicitly kill the bubblewrap child process,
when something goes wrong.