Add Wayland proxy support
This adds support for proxing the Wayland socket via
wayland-proxy-virtwl. It limits the supported Wayland protocols
to a sane set and allows prefixing window titles.
Although wayland-proxy-virtwl currently does not fully validate the
requests relayed to the Wayland compositor, it is already useful.
Furthermore, it looks like it might get used in future Qubes OS
releases for safeguarding the compositor. See:
https://github.com/talex5/wayland-proxy-virtwl/issues/90
https://github.com/talex5/ocaml-wayland/pull/49