Add PipeWire support
This adds PipeWire support by using pw-container to create a
separate security context and supplying the properties used by
Snap's PipeWire access control to limit playback and recording.

This is not a proxy solution like DBus or Wayland, but only uses
the security context feature integrated into PipeWire.
Security-wise this is not optimal, since the sandboxed applications
still directly talk to the PipeWire server. Furthermore, according
to Qubes OS developers PipeWire is currently not safe for
untrusted clients.
However, it is not worse than directly passing the default PipeWire
socket to applications and at least some work on securing PipeWire
is ongoing:

* https://gitlab.freedesktop.org/pipewire/pipewire/-/merge_requests/1384
* https://gitlab.freedesktop.org/pipewire/pipewire/-/merge_requests/1385
* https://gitlab.freedesktop.org/pipewire/pipewire/-/merge_requests/1820
* https://gitlab.freedesktop.org/pipewire/pipewire/-/merge_requests/2423
* https://gitlab.freedesktop.org/pipewire/pipewire/-/merge_requests/2442
* https://gitlab.freedesktop.org/pipewire/pipewire/-/merge_requests/2443
* https://gitlab.freedesktop.org/pipewire/pipewire/-/merge_requests/2446
* https://gitlab.freedesktop.org/pipewire/pipewire/-/merge_requests/2447
* https://gitlab.freedesktop.org/pipewire/pipewire/-/merge_requests/2449
* https://gitlab.freedesktop.org/pipewire/pipewire/-/merge_requests/2450
* https://gitlab.freedesktop.org/pipewire/pipewire/-/merge_requests/2453
* https://gitlab.freedesktop.org/pipewire/pipewire/-/merge_requests/2455